api/requirements.txt
Flask==2.2.2
openai==0.27.2
requests==2.28.1
pyyaml==6.0
aiohttp==3.8.3
langdetect==1.0.9